unify dcap types
