+++ crates/sgx-attestation/src/dcap.rs
@@ -12,7 +12,7 @@ use alloc::vec::Vec;
 use scale::{Decode, Encode};
 use scale_info::TypeInfo;
 
-use crate::dcap::quote::{AttestationKeyType, EnclaveReport, Quote, QuoteAuthData, QuoteVersion};
+use crate::dcap::quote::{AuthData, EnclaveReport, Quote};
 use crate::dcap::tcb_info::TcbInfo;
 use crate::dcap::utils::*;
 use crate::dcap::constants::*;
@@ -39,7 +39,8 @@ pub fn verify(
 ) -> Result<([u8; 64], Vec<u8>, String, Vec<String>), Error> {
     // Parse data
 
-    let quote = Quote::parse(raw_quote).map_err(|_| Error::CodecError)?;
+    let mut quote = raw_quote;
+    let quote = Quote::decode(&mut quote).map_err(|_| Error::CodecError)?;
 
     let tcb_info = pink_json::from_str::<TcbInfo>(&quote_collateral.tcb_info)
         .map_err(|_| Error::CodecError)?;
@@ -86,75 +87,74 @@ pub fn verify(
     }
 
     // Check quote fields
-    if quote.header.version != QuoteVersion::V3 {
+    if quote.header.version != QUOTE_VERSION_V3 {
         return Err(Error::UnsupportedDCAPQuoteVersion);
     }
     // We only support ECDSA256 with P256 curve
-    if quote.header.attestation_key_type != AttestationKeyType::ECDSA256WithP256Curve {
+    if quote.header.attestation_key_type != ATTESTATION_KEY_TYPE_ECDSA256_WITH_P256_CURVE {
         return Err(Error::UnsupportedDCAPAttestationKeyType);
     }
 
     // Extract Auth data from quote
-    let QuoteAuthData::Ecdsa256Bit {
-        signature,
-        attestation_key,
-        qe_report,
-        qe_report_signature,
-        qe_auth_data,
-        certification_data,
-    } = quote.signed_data
-    else {
+    let AuthData::V3(auth_data) = quote.auth_data else {
         return Err(Error::UnsupportedQuoteAuthData);
     };
+    let certification_data = auth_data.certification_data;
 
     // We only support 5 -Concatenated PCK Cert Chain (PEM formatted).
-    if certification_data.data_type != 5 {
+    if certification_data.cert_type != 5 {
         return Err(Error::UnsupportedDCAPPckCertFormat);
+    }
+
+    let certification_certs = extract_certs(&certification_data.body.data)?;
+    if certification_certs.len() < 2 {
+        return Err(Error::CertificateChainIsTooShort);
     }
     // Check certification_data
     let leaf_cert: webpki::EndEntityCert =
-        webpki::EndEntityCert::try_from(&certification_data.certs[0])
+        webpki::EndEntityCert::try_from(&certification_certs[0])
             .map_err(|_| Error::LeafCertificateParsingError)?;
-    let intermediate_certs = &certification_data.certs[1..];
+    let intermediate_certs = &certification_certs[1..];
     verify_certificate_chain(&leaf_cert, intermediate_certs, now_in_milli)?;
 
     // Check QE signature
-    let asn1_signature = encode_as_der(&qe_report_signature)?;
+    let asn1_signature = encode_as_der(&auth_data.qe_report_signature)?;
     if leaf_cert
-        .verify_signature(webpki::ECDSA_P256_SHA256, &qe_report, &asn1_signature)
+        .verify_signature(webpki::ECDSA_P256_SHA256, &auth_data.qe_report, &asn1_signature)
         .is_err()
     {
         return Err(Error::RsaSignatureIsInvalid);
     }
 
     // Extract QE report from quote
-    let parsed_qe_report =
-        EnclaveReport::from_slice(&qe_report).map_err(|_err| Error::CodecError)?;
+    let mut qe_report = auth_data.qe_report.as_slice();
+    let qe_report =
+        EnclaveReport::decode(&mut qe_report).map_err(|_err| Error::CodecError)?;
 
     // Check QE hash
     let mut qe_hash_data = [0u8; QE_HASH_DATA_BYTE_LEN];
-    qe_hash_data[0..ATTESTATION_KEY_LEN].copy_from_slice(&attestation_key);
-    qe_hash_data[ATTESTATION_KEY_LEN..].copy_from_slice(&qe_auth_data);
+    qe_hash_data[0..ATTESTATION_KEY_LEN].copy_from_slice(&auth_data.ecdsa_attestation_key);
+    qe_hash_data[ATTESTATION_KEY_LEN..].copy_from_slice(&auth_data.qe_auth_data.data);
     let qe_hash = ring::digest::digest(&ring::digest::SHA256, &qe_hash_data);
-    if qe_hash.as_ref() != &parsed_qe_report.report_data[0..32] {
+    if qe_hash.as_ref() != &qe_report.report_data[0..32] {
         return Err(Error::QEReportHashMismatch);
     }
 
     // Check signature from auth data
     let mut pub_key = [0x04u8; 65]; //Prepend 0x04 to specify uncompressed format
-    pub_key[1..].copy_from_slice(&attestation_key);
+    pub_key[1..].copy_from_slice(&auth_data.ecdsa_attestation_key);
     let peer_public_key =
         ring::signature::UnparsedPublicKey::new(&ring::signature::ECDSA_P256_SHA256_FIXED, pub_key);
     peer_public_key
         .verify(
             &raw_quote[..(HEADER_BYTE_LEN + ENCLAVE_REPORT_BYTE_LEN)],
-            &signature,
+            &auth_data.ecdsa_signature,
         )
         .map_err(|_| Error::IsvEnclaveReportSignatureIsInvalid)?;
 
     // Extract information from the quote
 
-    let extension_section = get_intel_extension(&certification_data.certs[0])?;
+    let extension_section = get_intel_extension(&certification_certs[0])?;
     let cpu_svn = get_cpu_svn(&extension_section)?;
     let pce_svn = get_pce_svn(&extension_section)?;
     let fmspc = get_fmspc(&extension_section)?;
@@ -187,13 +187,13 @@ pub fn verify(
     }
 
     let mut pruntime_hash = Vec::new();
-    pruntime_hash.extend_from_slice(&quote.enclave_report.mr_enclave);
-    pruntime_hash.extend_from_slice(&quote.enclave_report.isv_prod_id.to_be_bytes());
-    pruntime_hash.extend_from_slice(&quote.enclave_report.isv_svn.to_be_bytes());
-    pruntime_hash.extend_from_slice(&quote.enclave_report.mr_signer);
+    pruntime_hash.extend_from_slice(&quote.report.mr_enclave);
+    pruntime_hash.extend_from_slice(&quote.report.isv_prod_id.to_be_bytes());
+    pruntime_hash.extend_from_slice(&quote.report.isv_svn.to_be_bytes());
+    pruntime_hash.extend_from_slice(&quote.report.mr_signer);
 
     Ok((
-        quote.enclave_report.report_data,
+        quote.report.report_data,
         pruntime_hash,
         tcb_status.to_string(),
         advisory_ids,
